messaging-api/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Kanisorn-S <karamahati@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
line-bot-sdk = "^3.13.0"
uvicorn = "^0.32.0"
paho-mqtt = "^2.1.0"
fastapi-mqtt = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
